+++ example_test.go
@@ -2,6 +2,7 @@ package xr
 
 import (
 	"fmt"
+	"log"
 	"net/http"
 	"net/http/httptest"
 	"strings"
@@ -11,17 +12,10 @@ func ExamplePick_default() {
 	// handler on server side
 	h := func(w http.ResponseWriter, r *http.Request) {
 		var x PersonCreate
-		err := Pick(&x, r)
-		fmt.Println("id:", x.Id)
-		fmt.Println("name:", x.Name)
-		fmt.Println("group:", x.Group)
-		fmt.Println("copy:", x.Copy)
-		fmt.Println("flag:", x.Flag)
-		fmt.Println("pval:", x.PVal)
-		fmt.Println("token:", x.token)
-		fmt.Println("color:", x.Color)
-		fmt.Println("width:", x.Width)
-		fmt.Println(err)
+		if err := Pick(&x, r); err != nil {
+			log.Fatal(err)
+		}
+		fmt.Print(x)
 	}
 
 	w := httptest.NewRecorder()
@@ -36,16 +30,7 @@ func ExamplePick_default() {
 	mux.HandleFunc("/person/{id}", h)
 	mux.ServeHTTP(w, r)
 	// output:
-	// id: 123
-	// name: John Doe
-	// group: aliens
-	// copy: 10
-	// flag: true
-	// pval: 11.79
-	// token: ...token...
-	// color: yellow
-	// width: 100
-	// <nil>
+	// {123 John Doe aliens 10 true 11.79 Bearer ...token... yellow 100 }
 }
 
 type PersonCreate struct {
@@ -56,29 +41,10 @@ type PersonCreate struct {
 	Flag  bool    `query:"flag"`
 	PVal  float32 `query:"pval"`
 
-	// private field requires method SetToken
-	token string `header:"authorization"`
-
 	Auth  string `header:"authorization"`
 	Color string `header:"Color"`
-	Width int    `json:"width" minimum:"200"`
-}
+	Width int    `json:"width"`
 
-// SetToken trims optional prefix "Bearer " from v.
-func (p *PersonCreate) SetToken(v string) {
-	if strings.HasPrefix(v, "Bearer ") {
-		p.token = v[7:]
-		return
-	}
-	p.token = v
-}
-
-func (p *PersonCreate) SetColor(v string) error {
-	switch v {
-	case "black":
-		return fmt.Errorf("color unsupported: %s", v)
-	default:
-		p.Color = v
-	}
-	return nil
+	// private fields are ignored
+	token string
 }